nvim: inline gitsigns config
Also remove dependencies from snacks.lua that are not need at config
time, which resolved a loading order bug. Gitsigns and copilot pluging
are used in the config of snacks, but a require("…") does not happen
at config time, it only sets up handlers that do that. So, they should
not be in the dependencies list.

The config of gitsigns however has to require("snacks.picker") and
therefore snacks has to go in the dependencies list of gitsigns, so that
it is loaded before gitsigns and the require does not fail.